# SequenceReader performance; elide bounds checks and book-keeping by using ref-oriented approach