--- a/src/libraries/System.Memory/src/System/Buffers/SequenceReader.Search.cs
+++ b/src/libraries/System.Memory/src/System/Buffers/SequenceReader.Search.cs
@@ -831,7 +831,7 @@ namespace System.Buffers
                 next = next.Slice(currentSpan.Length);
 
                 // Need to check the next segment
-                if (!TryGetNextBuffer(in _sequence, ref segment, ref currentSpan, out _))
+                if (TryGetNextBuffer(in _sequence, ref segment, ref currentSpan) != TryGetNextBufferResult.SuccessHaveData)
                 {
                     // Nothing left
                     return false;