… src/libraries/System.Memory/src/System/Buffers/SequenceReader.cs