context: https://twitter.com/marcgravell/status/1629955240895078405 (read back for @davidfowl mention)

purpose: improved performance in `SequenceReader`

approach:

- reduced book-keeping by usually only manipulating a single index
- remove bounds-checking (since we're checking the bounds ourselves) via `Unsafe.AddRef`
- when possible (net7+) store the current span's `ref T` directly