@mgravell 2023-02-28
@stephentoub well, I won't pretend not to be disappointed - however, thoughts on just reducing the book-keeping etc?