@stephentoub 2023-03-01
> thoughts on just reducing the book-keeping etc?

If there are safe ways of making it faster/smaller/etc., that's great, let's do it.  And as I noted, I expect there are ways of writing the code differently such that bounds checks can be eliminated even without using Unsafe.