@mgravell 2023-03-01
disclosure: I'm not currently able to run the tests locally (although it builds cleanly):

> Failed to launch testhost with error: System.IO.FileNotFoundException: Path does not exist: C:\Code\runtime\sequencereader\artifacts\bin\testhost\net8.0-windows-Debug-x64\dotnet.exe
File name: 'C:\Code\runtime\sequencereader\artifacts\bin\testhost\net8.0-windows-Debug-x64\dotnet.exe'

I've tried about 30 things, but I can't get that working right now, and I can't find an idiot-proof guide to making it happy.