@stephentoub 2023-03-01
> disclosure: I'm not currently able to run the tests locally (although it builds cleanly):

How are you building and running tests?