@mgravell 2023-03-01
locally, with .net8 preview installed; at command-line, `dotnet build` in `...\src\libraries\System.Memory\src` and and `...\src\libraries\System.Memory\tests` - and `dotnet test` in `...\src\libraries\System.Memory\tests` - fails as above; also using IDE via `\src\libraries\System.Memory\System.Memory.sln`, but fails with same error in output window

in root, `build.cmd` works in my `main` branch, but does not generate the missing artifacts; oddly `build.cmd` *does not* work in my branch, failing with a build.ninja failure about a missing C compiler; which is weird because it works everywhere else, including the coreclr build