@mgravell 2023-03-01
done all that (using `/t:test` - `tests` wasn't found); no change - it is still trying to use `dotnet` from `...\artifacts\bin\testhost\net8.0-windows-Debug-x64`, which is empty, regardless of `dotnet` vs `dotnet.cmd`; the `dotnet.cmd` from the root seems to be using the program-files version of preview1, so... yeah, no idea what is going on; just:

> '"C:\Code\runtime\sequencereader\artifacts\bin\testhost\net8.0-windows-Debug-x64\dotnet.exe"' is not recognized as an internal or external command, operable program or batch file.

(because it does not exist)