@mgravell 2023-03-02
interestingly, `TryPeek` seems to suffer a weird regression **only** on net7 (fine on net6 and net8); I propose to ignore net7, and I'll re-run the benchmarks only on net6 and net8 (I'm very intrigued if anyone knows what this regression may be - presumably a JIT change around branches):

```
        [MethodImpl(MethodImplOptions.AggressiveInlining)]
        public readonly bool TryPeek(out T value) {
            AssertValidPosition(); // note: [Conditional("DEBUG")]
            if (_currentSpanIndex == _currentSpan.Length) // only true at EOF due to eager read
            {
                value = default;
                return false;
            }
            value = _currentSpan[_currentSpanIndex];
            return true;
        }
```

results:

``` txt
|        Method |      Job |  Runtime | SegmentLength |       Mean |    Error |   StdDev |
|-------------- |--------- |--------- |-------------- |-----------:|---------:|---------:|
| SystemTryPeek | .NET 6.0 | .NET 6.0 |         -1024 |   747.2 ns | 14.42 ns | 14.81 ns |
| CustomTryPeek | .NET 6.0 | .NET 6.0 |         -1024 |   687.9 ns |  9.98 ns |  9.33 ns |
| SystemTryPeek | .NET 6.0 | .NET 6.0 |             4 |   749.9 ns | 11.27 ns |  9.99 ns |
| CustomTryPeek | .NET 6.0 | .NET 6.0 |             4 |   692.4 ns | 11.89 ns | 10.54 ns |
| SystemTryPeek | .NET 6.0 | .NET 6.0 |           128 |   742.4 ns | 10.89 ns | 10.19 ns |
| CustomTryPeek | .NET 6.0 | .NET 6.0 |           128 |   690.3 ns |  9.60 ns |  8.98 ns |
| SystemTryPeek | .NET 6.0 | .NET 6.0 |          1024 |   743.5 ns |  5.95 ns |  5.56 ns |
| CustomTryPeek | .NET 6.0 | .NET 6.0 |          1024 |   682.0 ns |  3.82 ns |  3.57 ns |
|               |          |          |               |            |          |          |
| SystemTryPeek | .NET 7.0 | .NET 7.0 |         -1024 |   679.6 ns |  8.15 ns |  7.23 ns |
| CustomTryPeek | .NET 7.0 | .NET 7.0 |         -1024 | 1,121.2 ns | 12.48 ns |  9.74 ns |
| SystemTryPeek | .NET 7.0 | .NET 7.0 |             4 |   683.3 ns | 11.10 ns | 10.38 ns |
| CustomTryPeek | .NET 7.0 | .NET 7.0 |             4 | 1,123.8 ns |  9.98 ns |  8.33 ns |
| SystemTryPeek | .NET 7.0 | .NET 7.0 |           128 |   685.8 ns |  9.32 ns |  8.72 ns |
| CustomTryPeek | .NET 7.0 | .NET 7.0 |           128 | 1,118.5 ns | 10.64 ns |  9.95 ns |
| SystemTryPeek | .NET 7.0 | .NET 7.0 |          1024 |   679.0 ns |  6.24 ns |  5.84 ns |
| CustomTryPeek | .NET 7.0 | .NET 7.0 |          1024 | 1,116.6 ns |  6.57 ns |  5.49 ns |
|               |          |          |               |            |          |          |
| SystemTryPeek | .NET 8.0 | .NET 8.0 |         -1024 |   704.7 ns | 13.94 ns | 25.50 ns |
| CustomTryPeek | .NET 8.0 | .NET 8.0 |         -1024 |   630.5 ns |  8.47 ns |  7.92 ns |
| SystemTryPeek | .NET 8.0 | .NET 8.0 |             4 |   690.5 ns |  9.48 ns |  8.86 ns |
| CustomTryPeek | .NET 8.0 | .NET 8.0 |             4 |   625.4 ns |  5.49 ns |  4.59 ns |
| SystemTryPeek | .NET 8.0 | .NET 8.0 |           128 |   686.4 ns |  8.86 ns |  8.29 ns |
| CustomTryPeek | .NET 8.0 | .NET 8.0 |           128 |   624.5 ns |  5.68 ns |  5.31 ns |
| SystemTryPeek | .NET 8.0 | .NET 8.0 |          1024 |   696.3 ns | 12.07 ns | 10.70 ns |
| CustomTryPeek | .NET 8.0 | .NET 8.0 |          1024 |   631.1 ns |  8.28 ns |  7.74 ns |
```