@stephentoub 2023-03-02
I'm not sure why there's a .NET 7 regression, but regardless of target, I suggest trying pulling the fields into locals prior to doing the comparison; you can do it in a way that'll result in smaller code and no bounds check:
[SharpLab](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8BiAOwFcAbS7YSmAAgBMBLXW+gWAChiAHKNgDm+bAwDu2KKRakhzNh0YAGVAE4eAAQBMARi3KGm3QDoASuVIYW+GCYDCEfHxb0oAZRhQAbizAxcAG4DBncACyk+ABlac0trW2DuHgBtACkWDABxGFIvPwAKDABPPhgIADMC2QwASlqAXS0AZgZYCoZcDChyMAwGewAeABUAPh4Abx4GGYYBFm9sDEYahgB9MHIoWCt3PmxSAElSJhgEJNm5qAWlxjMYbCYAeVJKYr2DkdH1ze3cjA+pAus2msxSAFkYBgwhAmIdnJQCpDobD4XxKE8+NYIKRcCYAIJCISwXC4BYwY6UWSyISNUEzTStWCPHFvBjACAQSgMYZQYoABRgMAA1roChByP1hgxFpRyDBavSGFNuJdLiwOgUNlsdgD9kcTmcGABeY0/HX/QEmKK5ITQxWqtUzFVOp2y+Um5gwCrYKgYYGu2aaADsDB9lFwMADaoAvkrLu7GGbtX9dvqUindYDjqcEA1o5cQwxuvLo3HHTMlRCoTC4QikTXUQjMdjcQSiSSyd4Ka9qXI6RWjEyHkxWcV2ZzubyBULhdpxZKeTLsHKFUqXa7VixPZnLfqc2cC7N7o8Xm9AV8GL8s/qd9e9wcj5XB+rNQVyDVatvBgx35/72mBzWra9rxrMG6BjMiaegBeoHCkLD5mBTpFiWUbIQw5aBhh0FmqcPp+k+RihuGkZljwMZAA=)