@Jozkee 2023-08-14
I managed to rebase this PR on top of main and compared the results against main. There are very nice improvements with the exception of `TryReadTo` which shows a regression, specially on the small segment length case. We may want to hold-off this change until 9.0.

<details>
  <summary>Results</summary>

```
BenchmarkDotNet v0.13.7-nightly.20230717.35, Windows 11 (10.0.22621.2134/22H2/2022Update/SunValley2)
AMD Ryzen 9 5900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK 8.0.100-preview.7.23376.3
  [Host]     : .NET 8.0.0 (8.0.23.37506), X64 RyuJIT AVX2
  Job-VDEYSP : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  Job-JFPICV : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:EnableUnsafeBinaryFormatterSerialization=true  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  


|               Method |        Job |                                                                                                  Toolchain | SegmentLength |        Mean |     Error |    StdDev |      Median |         Min |         Max | Ratio | MannWhitney(3ms) | RatioSD | Allocated | Alloc Ratio |
|--------------------- |----------- |----------------------------------------------------------------------------------------------------------- |-------------- |------------:|----------:|----------:|------------:|------------:|------------:|------:|----------------- |--------:|----------:|------------:|
| SystemAdvancePastAny | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   235.90 ns |  2.218 ns |  1.852 ns |   236.46 ns |   232.99 ns |   238.75 ns |  0.55 |             Same |    0.01 |         - |          NA |
| SystemAdvancePastAny | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   432.11 ns |  2.496 ns |  2.335 ns |   432.02 ns |   428.73 ns |   436.98 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
| SystemAdvancePastAny | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 | 1,675.89 ns |  4.426 ns |  3.923 ns | 1,674.76 ns | 1,671.46 ns | 1,684.04 ns |  0.94 |             Same |    0.00 |         - |          NA |
| SystemAdvancePastAny | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 | 1,785.57 ns |  7.547 ns |  6.690 ns | 1,785.14 ns | 1,775.78 ns | 1,798.89 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
| SystemAdvancePastAny | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   313.67 ns |  3.434 ns |  3.212 ns |   313.32 ns |   310.05 ns |   320.11 ns |  0.62 |             Same |    0.01 |         - |          NA |
| SystemAdvancePastAny | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   507.31 ns |  1.234 ns |  1.093 ns |   507.05 ns |   506.04 ns |   509.61 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
| SystemAdvancePastAny | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   233.57 ns |  1.798 ns |  1.682 ns |   232.56 ns |   232.03 ns |   237.22 ns |  0.54 |             Same |    0.00 |         - |          NA |
| SystemAdvancePastAny | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   435.17 ns |  0.879 ns |  0.822 ns |   434.92 ns |   434.27 ns |   436.68 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemAdvance | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   442.82 ns |  6.444 ns |  6.027 ns |   441.80 ns |   436.09 ns |   453.74 ns |  1.00 |             Same |    0.01 |         - |          NA |
|        SystemAdvance | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   440.75 ns |  1.499 ns |  1.171 ns |   440.77 ns |   438.91 ns |   442.45 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemAdvance | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 | 1,939.14 ns |  6.711 ns |  6.278 ns | 1,938.32 ns | 1,930.90 ns | 1,949.90 ns |  0.97 |             Same |    0.01 |         - |          NA |
|        SystemAdvance | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 | 1,992.90 ns |  6.847 ns |  6.070 ns | 1,993.70 ns | 1,985.12 ns | 2,006.13 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemAdvance | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   514.84 ns |  1.342 ns |  1.190 ns |   514.68 ns |   513.30 ns |   517.49 ns |  0.96 |             Same |    0.00 |         - |          NA |
|        SystemAdvance | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   537.23 ns |  1.004 ns |  0.839 ns |   537.03 ns |   536.11 ns |   538.77 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemAdvance | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   439.22 ns |  1.397 ns |  1.238 ns |   438.82 ns |   437.83 ns |   441.31 ns |  0.99 |             Same |    0.00 |         - |          NA |
|        SystemAdvance | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   443.68 ns |  1.115 ns |  1.043 ns |   443.44 ns |   442.62 ns |   445.81 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|       SystemPosition | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   224.71 ns |  0.797 ns |  0.746 ns |   224.38 ns |   224.01 ns |   226.05 ns |  0.33 |             Same |    0.00 |         - |          NA |
|       SystemPosition | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   681.10 ns |  2.445 ns |  2.287 ns |   681.14 ns |   677.98 ns |   686.19 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|       SystemPosition | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 |   450.55 ns |  1.117 ns |  1.045 ns |   450.19 ns |   449.03 ns |   452.32 ns |  0.25 |             Same |    0.00 |         - |          NA |
|       SystemPosition | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 | 1,772.73 ns | 14.692 ns | 13.743 ns | 1,779.07 ns | 1,745.56 ns | 1,790.50 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|       SystemPosition | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   227.25 ns |  0.687 ns |  0.643 ns |   226.90 ns |   226.65 ns |   228.37 ns |  0.13 |             Same |    0.00 |         - |          NA |
|       SystemPosition | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 | 1,728.64 ns |  3.115 ns |  2.914 ns | 1,728.06 ns | 1,725.46 ns | 1,734.62 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|       SystemPosition | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   227.99 ns |  0.721 ns |  0.675 ns |   227.68 ns |   227.40 ns |   229.30 ns |  0.33 |             Same |    0.00 |         - |          NA |
|       SystemPosition | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   682.82 ns |  2.620 ns |  2.451 ns |   682.41 ns |   679.96 ns |   687.86 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemTryPeek | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   434.19 ns |  0.631 ns |  0.527 ns |   434.01 ns |   433.71 ns |   435.55 ns |  0.95 |             Same |    0.01 |         - |          NA |
|        SystemTryPeek | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |   456.50 ns |  8.852 ns |  8.694 ns |   454.83 ns |   448.30 ns |   478.09 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemTryPeek | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 |   443.23 ns |  2.940 ns |  2.606 ns |   442.49 ns |   440.53 ns |   448.91 ns |  0.99 |             Same |    0.02 |         - |          NA |
|        SystemTryPeek | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 |   449.71 ns |  7.907 ns |  7.396 ns |   447.17 ns |   440.26 ns |   462.55 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemTryPeek | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   444.70 ns |  2.642 ns |  2.206 ns |   444.63 ns |   441.52 ns |   448.45 ns |  0.98 |             Same |    0.01 |         - |          NA |
|        SystemTryPeek | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   452.82 ns |  5.928 ns |  5.545 ns |   453.32 ns |   442.99 ns |   462.30 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|        SystemTryPeek | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   438.19 ns |  1.099 ns |  1.028 ns |   438.31 ns |   436.65 ns |   439.89 ns |  0.94 |             Same |    0.01 |         - |          NA |
|        SystemTryPeek | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |   468.28 ns |  7.365 ns |  6.889 ns |   471.56 ns |   459.28 ns |   478.39 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|      SystemTryReadTo | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |    85.16 ns |  0.258 ns |  0.241 ns |    85.06 ns |    84.88 ns |    85.60 ns |  1.04 |             Same |    0.00 |         - |          NA |
|      SystemTryReadTo | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |         -1024 |    82.04 ns |  0.482 ns |  0.451 ns |    82.10 ns |    80.68 ns |    82.57 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|      SystemTryReadTo | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 | 3,482.61 ns | 12.050 ns | 11.272 ns | 3,481.28 ns | 3,461.31 ns | 3,504.64 ns |  1.82 |             Same |    0.00 |         - |          NA |
|      SystemTryReadTo | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |             4 | 1,913.75 ns |  4.098 ns |  3.633 ns | 1,911.78 ns | 1,910.57 ns | 1,921.80 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|      SystemTryReadTo | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   168.54 ns |  0.508 ns |  0.450 ns |   168.56 ns |   167.83 ns |   169.15 ns |  1.33 |             Same |    0.02 |         - |          NA |
|      SystemTryReadTo | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |           128 |   126.69 ns |  1.687 ns |  1.578 ns |   126.93 ns |   123.68 ns |   129.13 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
|      SystemTryReadTo | Job-VDEYSP | \runtime2\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |    87.80 ns |  0.252 ns |  0.236 ns |    87.79 ns |    87.50 ns |    88.17 ns |  1.36 |             Same |    0.01 |         - |          NA |
|      SystemTryReadTo | Job-JFPICV |  \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |          1024 |    64.73 ns |  0.420 ns |  0.392 ns |    64.69 ns |    64.24 ns |    65.47 ns |  1.00 |             Base |    0.00 |         - |          NA |
|                      |            |                                                                                                            |               |             |           |           |             |             |             |       |                  |         |           |             |
…